requirements.txt
Flask
SQLAlchemy
PyMySQL
python-dotenv
bcrypt
passlib[bcrypt]
gunicorn